Add other ops
